requirements.txt
langchain
langchain_cohere
langgraph
streamlit
python-dotenv
mysql-connector-python